Improve computation of text vertical dimension
Some fonts possess glyphs which ascents/descents are larger that the font
global ascent/descent. This is known to happen on macOS.

This patch introduces code to make these individual ascents/descents available:
1/ FontMetrics::breakString now provides a full dimension for each line
   of text that it produces
2/ The new FontMetrics::dimension provides the dimension of a given string
3/ FontMetrics::rectText is modified to rely on this new method.

The new code is used when constructing rows to replace some of the
uses of FontMetrics::maxAscent/Descent by proper values. This holds in
particular for calculation of row height.

Possible fix to bug #12879.