Do not draw appendix lines below en of document
This bug is pretty old, but it was not visible because the grey area
below the document would overwrite it.

It would be better to close the frame at the end of the document, but
this is for later.

(cherry picked from commit 7acfbe0fccc729dcf5d910a4049b25e9943329fc)
(cherry picked from commit c96d3a03400893f440dca7e68ed87bd70bea0b39)