ctests: invert chess lyx2lyx tests
In export.cmake, if the detected output format is pdf3, we try to
export to "platex", which usually makes sense, but for this file we
would need to export to "latex". This file is very particular, and
also old, so I do not know if it is worth the time to handle this
case.

This commit inverts the following tests:

  export/examples/Articles/Chess/Game_1_lyx16 (Failed)
  export/examples/Articles/Chess/Game_1_lyx20 (Failed)
  export/examples/Articles/Chess/Game_1_lyx21 (Failed)
  export/examples/Articles/Chess/Game_1_lyx22 (Failed)
  export/examples/Articles/Chess/Game_1_lyx23 (Failed)