Amend 7feffb89: restore the meaning of the elements in the QComboBox.
The order is not user-friendly, but correct. Before this patch, the elements in the UI didn't correspond to the implemented behaviour. The problem is the order: the first element is always MathML, the second HTML, etc. I think we should rather use some `userData` in QComboBox items to decouple the order from the semantics.