Fixup 5e857883: update screen when clicking away from preview
This is a regression in 5e857883 that comes from the fact that the
Update::singlePar flag does not operate on the paragraph that may be
modified but on the one where the cursor currently lies.

Until this is fixed, we do a full update when this situation is encountered.

Fixes bug #13128.