Fix typo that make the conversion not working
The reason to change was that format is a builtin function in Python.

By redefining it we are simply overriding it and so it is not a problem
although not elegant.

Another issue in this file is `match` that become a contextual keyword
in Python 3.10. Again it is not elegant although it is allowed.