Ask user to save changes before closing preamble editor (#8145)
This prevents unexpected data loss when the document dialog is closed
with either the ESC key or the cross button and there are unapplied
changes (preamble or elsewhere). Similar to what we already do when
the buffer is closed on these conditions.

(cherry picked from commit 6acc0759fc89e1d309a7f24dddee7ddac4226550)