Introducing OnTheSpot-style input method editing & IME integration
OnTheSpot style of the input method editor (IME) shows preedit strings in the
same status as normal string inputs instead of showing it overlapping with
the main text. In doing so, the input method is more integrated with LyX
enabling e.g. automatic on/off of IME in the math editor or in the command mode.