Amend 9b22c347
Better change the name of the configure option rather than
clearing the name of the cached variable before the check.

Indeed, this would defeat the reason for caching, even if in
our case it makes a difference only if persistent caching
is activated by using the -C or --config-cache option to configure.

However, note that persistent caching is not recommended:
https://autotools.info/autoconf/caching.html