InsetERT: recognise whenever the contents looks sufficiently like XML to attempt outputting it as raw XML.
This behaviour is closer to what LyX 2.3 and previous handled ERTs for DocBook: they were always output as SGML. This no longer makes sense for the new DocBook support (you can export basically any document to DocBook and have reasonable output, while previously you had to use a DocBook template), but this heuristic should capture most previous use cases. (If the ERT is just XML: output it. Otherwise, it's probably not meant to be output as-is -- LaTeX, text, etc.)