Make a test for Cocoa dynamic
In general, it is a good idea to replace when possible #ifdef that
rest for a platform with calls to QGuiQpplication::platformName. This
allow the same application to run for example with the xcb backend (I
am not sure this will be possible one day with macOS, though).

As often these days, the real motivation for changing this was to calm
down Coverity Scan.