Use while() loops instead of fighting against for()
In a loop where elements may be removed from a vector, it is dangerous
and weird to decrease the counter so that it can be incremented by the
for() ++i statement later without change. I suspect that things can go wrong
when i == 0, and at least Coverity Scan agrees.

Use while() instead.