A bit more color to the world, part I
In addition to the rather small list of mostly unpleasant "base colors"
as defined by color and xcolor, we now support the whole range of
(a whopping 752) named colors provided by the xcolor package.

The colors are only added to LyX's color cache and list if actually
used, but they are all accessible (for math and text) via the char
style dialog (and we'll see if the UI needs improvement), and known
to tex2lyx.

In forthcoming parts, these colors will be made accessible also in
other places (e.g., boxes), and I will add the possibility to add to
the list own custom colors (as requested in #1841).

NOTE that this will drop on you a frightening number of new GUI strings,
many of which, however, (the more technical color names) should probably
simply be taken over from the original (English) form for l7n.