Fixup 16e67d4e: replace incorrect code with with new mechanism
The commit above tried to avoid recursion in InsetMathMacro::validate
by readong macro definitions. Unfortunately, the code is broken (for
example tokenPos does not return an index in a string) and Coverity
Scan does not like it (because the pos variable can be used although
it is negative).

This is a fresh start: instead of avoiding calling validate, we return
early when the recursion happens. To this end, we use a static
set<docstring> that checks what macros are currently being validated.
This still works when the recursion happens in a cycle of size 3 or
more.

Related to bug #12633.