Use _ convention for unused variable
The usual convention, also recognised by linters, is to use an
underscore when we explicitly want to ignore one function output.

The other option would be not assign any value to the function output,
but in this case we can not distinguish if this was done on purpose or
not.