lyx-tester switches: delay calculations
This file is typically sourced by the user into the bash sessions,
but created these contsants takes about half a second, which the
user does not want when not calling the switches.

Instead, the calculations are still centralized but now implemented
as functions so that the calculations are done only when used.