Select a branch after rm a branch in doc settings
Before, after pressing the "Remove" PB to remove a branch, there
would be no subsequent branch selected. Now, if there is a branch
below the one removed, it is selected; if not, if there is a branch
above the one removed, it is selected.

This is useful for removing multiple branches by clicking several
times on the "Remove" PB.