Debug msg if adding existing branch in container
It is probably better to handle these cases at a higher level, as
was just done at 7d1eea03.

I will not be surprised if there is a case (that I am not currently
aware of) where it is actually ideal to handle this case in this
low-level add() method.

For now, this reports an error as a debug message. Eventually I
would like to change this to an assert.