Let's keep track of the analysis.
It's much easier to keep track of the changes through the diffs,
in the short-mid term. Some of changes might not get detected
though so in a very long term the inspection needs to be done anew.

In the first iteration we just check .h files. List from older
codbase (fall 2024).